edit menu images

+++ app/Http/Controllers/api/ItemController.php
@@ -4,11 +4,9 @@ namespace App\Http\Controllers\api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Item;
-use App\Models\Order;
 use Illuminate\Http\Request;
 use App\Http\Resources\ItemResource;
 use App\Models\ItemAddition;
-use App\Models\OrderItem;
 use Illuminate\Support\Facades\Validator;
 class ItemController extends Controller
 {
@@ -25,11 +23,9 @@ class ItemController extends Controller
     {
         $item = Item::all();
         return ItemResource::collection($item);
+
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -44,7 +40,20 @@ class ItemController extends Controller
             return response($validator->errors()->all(), 422);
         }
 
-        $item = Item::create($request->all());
+        
+        $path = public_path('images/item');
+        !is_dir($path) &&
+            mkdir($path, 7777, true);
+
+
+        $imageName = time() . '.' . $request->file('image')->extension();
+        $request->file('image')->move($path, $imageName);
+
+        $fullRequest = $request->all();
+        $fullRequest['image'] = $imageName;
+        
+
+        $item = Item::create($fullRequest);
         $itemAdditionsArray = json_decode($request->additions, true);
         $this->storeItemAdditions($itemAdditionsArray, $item->id);
 
@@ -66,7 +75,8 @@ class ItemController extends Controller
      */
     public function show(Item $item)
     {
-        return new ItemResource($item);
+        return new ItemResource($item) ;
+
     }
 
     /**
@@ -74,23 +84,8 @@ class ItemController extends Controller
      */
     public function update(Request $request, Item $item)
     {
-        $validator = Validator::make($request->all(), [
-            "name" => "required",
-            "price" => "required",
-            "description" => "required",
-            "category_id" => "required",
-        ]);
-
-        if ($validator->fails()) {
-            return response($validator->errors()->all(), 422);
-        }
-
         $item->update($request->all());
-
-        ItemAddition::where('item_id', $item->id)->delete();
-        $this->storeItemAdditions($request->additions, $item->id);
-
-        return new ItemResource($item);
+        return new ItemResource ($item);
     }
 
     /**
@@ -98,10 +93,19 @@ class ItemController extends Controller
      */
     public function destroy(Item $item)
     {
-        $orders = Order::where('status', 'cart')->pluck('id');
-        OrderItem::whereIn('order_id', $orders)->where('item_id', $item->id)->delete();
 
-        $item->delete();
+        //  $product = Product::findorfail($id);
+        // $product->delete();
+$image_path=public_path('images/item/'. $item->img);
+if(file_exists($image_path)){
+    unlink($image_path);
+    $item->delete();
+
+}
+
+// dd($image_path);
+        // unlink('images/item/'. $item->img);
+
         return response("Deleted", 204);
     }
 }
